This commit enhances the whitelisting logic to reduce bugs caused by different assumptions between webworks and webkit whitelisting rules.
Now everytime a domain is whitelisted it automatically is given access
to and from file and local. This is a hack to account for the fact that
whitelisting a file path should have given EVERY domain access to it,
but we have no way to implement this in webkit.

It also adds a similair fix to give the content page access to and from
file and local when access is *. This is again a dirty hack, because in
webworks access * should give all domains access to file.